refactor: parse payload content in a new "content" module
also fix: were the inputs not being parsed in tests idkperhaps
